+++ YnabSharp.Seeder/Step3.Seed/Step33.SeedAccounts/SeedAccountsCliCommand.cs
@@ -3,4 +3,4 @@ using YnabSharp.Connected;
 
 namespace YnabSharp.Seeder.Step3.Seed.Step33.SeedAccounts;
 
-public record SeedAccountsCliCommand(ConnectedBudget Budget, List<Account> Accounts) : CliCommand;
+public record SeedAccountsCliCommand(ConnectedPlan Plan, List<Account> Accounts) : CliCommand;