+++ YnabSharp.Seeder/Step3.Seed/Step34.SeedTransactions/SeedTransactionsCliCommand.cs
@@ -3,4 +3,4 @@ using YnabSharp.Connected;
 
 namespace YnabSharp.Seeder.Step3.Seed.Step34.SeedTransactions;
 
-public record SeedTransactionsCliCommand(ConnectedBudget Budget, List<Transaction> Transactions) : CliCommand;
+public record SeedTransactionsCliCommand(ConnectedPlan Plan, List<Transaction> Transactions) : CliCommand;