+++ YnabSharp.Seeder/YnabSharpSeederRegistry.cs
@@ -1,7 +1,7 @@
 using KitCli.Abstractions;
 using KitCli.Commands.Abstractions.Extensions;
 using Microsoft.Extensions.DependencyInjection;
-using YnabSharp.Seeder.Step1.SetBudget;
+using YnabSharp.Seeder.Step1.SetPlan;
 
 namespace YnabSharp.Seeder;
 
@@ -9,6 +9,6 @@ public class YnabSharpSeederRegistry : ICliAppBuilderRegistry
 {
     public void Register(IServiceCollection services)
     {
-        services.AddCommandsFromAssembly(typeof(SetBudgetCliCommand).Assembly);
+        services.AddCommandsFromAssembly(typeof(SetPlanCliCommand).Assembly);
     }
 }