+++ YnabSharp/PlanYears.cs
@@ -1,11 +1,11 @@
 namespace YnabSharp;
 
-public class BudgetYears
+public class PlanYears
 {
     public List<int> All { get; set; }
     public List<int> Measurable { get; set; }
-    
-    public BudgetYears(List<int> all, List<int> measurable)
+
+    public PlanYears(List<int> all, List<int> measurable)
     {
         All = all;
         Measurable = measurable;